requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiohttp-retry==2.9.1
aiosignal==1.4.0
amqp==5.3.1
arabic-reshaper==3.0.0
arrow==1.4.0
asgiref==3.10.0
asn1crypto==1.5.1
attrs==25.4.0
autobahn==25.12.2
Automat==25.4.16
billiard==4.2.2
binaryornot==0.4.4
boto3==1.42.14
botocore==1.42.14
brotli==1.2.0
cbor2==5.8.0
celery==5.5.3
certifi==2025.10.5
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
chardet==5.2.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cookiecutter==2.6.0
crispy-bootstrap5==2025.6
cryptography==46.0.3
cssselect2==0.8.0
daphne==4.2.1
dj-database-url==3.0.1
Django==5.2.7
django-browser-reload==1.21.0
django-cors-headers==4.9.0
django-crispy-forms==2.4
django-debug-toolbar==6.0.0
django-phonenumber-field==8.3.0
django-ratelimit==4.1.0
django-redis==6.0.0
django-ses==4.5.0
django-storages==1.14.6
django-tailwind==4.4.1
django-widget-tweaks==1.5.0
djangorestframework==3.16.1
fonttools==4.60.1
freetype-py==2.5.1
frozenlist==1.8.0
gunicorn==23.0.0
honcho==2.0.0
html5lib==1.1
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
iniconfig==2.3.0
Jinja2==3.1.6
jmespath==1.0.1
kombu==5.5.4
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
msgpack==1.1.2
multidict==6.7.0
ordered-set==4.1.0
oscrypto==1.3.0
packaging==25.0
phonenumbers==9.0.17
pillow==12.0.0
pluggy==1.6.0
prompt_toolkit==3.0.52
propcache==0.4.1
psutil==7.2.1
psycopg2-binary==2.9.11
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.28.0
pycparser==2.23
pydyf==0.11.0
Pygments==2.19.2
pyHanko==0.31.0
pyhanko-certvalidator==0.29.0
PyJWT==2.10.1
pyOpenSSL==25.3.0
pypdf==6.2.0
pyphen==0.17.2
pytailwindcss==0.3.0
pytest==9.0.1
python-barcode==0.16.1
python-bidi==0.6.7
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
pytz==2025.2
PyYAML==6.0.3
redis==6.4.0
reportlab==4.4.4
requests==2.32.5
rich==14.2.0
rlPyCairo==0.4.0
s3transfer==0.16.0
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.3
svglib==1.6.0
text-unidecode==1.3
tinycss2==1.4.0
tinyhtml5==2.0.0
twilio==9.8.8
Twisted==25.5.0
txaio==25.12.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.3.1
ujson==5.11.0
uritools==5.0.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
weasyprint==66.0
webencodings==0.5.1
whitenoise==6.11.0
yarl==1.22.0
zope.interface==8.1.1
zopfli==0.4.0
